Exercises 4.8-4.14

+++ part4/BlogList/controllers/blogs.js
@@ -2,43 +2,56 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
-  respons.json(blogs)
+  const result = await Blog.find({})
+  response.json(result)
 })
 
-blogsRouter.get('/:id', (request, response, next) => {
-  Blog.findById(request.params.id)
-    .then(blog => {
-      if (blog) {
-        response.json(blog)
-      } else {
-        response.status(404).end()
-      }
-    })
-    .catch(error => next(error))
+blogsRouter.get('/:id', async (request, response, next) => {
+  const blog = await Blog.findById(request.params.id)
+  if (blog) {
+    response.json(blog.toJSON())
+  } else {
+    response.status(404).end()
+  }
 })
 
 blogsRouter.post('/', async (request, response, next) => {
-  const body = request.body
+  const blog = new Blog(request.body)
 
-  const blog = new Blog({
-    content: body.content,
-    important: body.important || false,
-    date: new Date()
-  })
-
-  const savedBlog = await blog.save()
-  response.json(savedBlog)
+if (!blog.likes){
+  blog.likes = 0
+}
+if (!blog.title && !blog.url) {
+  response
+    .status(400)
+    .json({
+      error: 'content missing',
+    })
+    .end()
+  return
+}
+  const result = await blog.save()
+  response.status(201).json(result)
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
-  try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
-  }catch (exception){
-    next(exception)
+blogsRouter.delete('/:id', async (request, response) => {
+  await Blog.findByIdAndRemove(request.params.id)
+  response.status(204).end()
+})
+
+blogsRouter.put('/:id', async (request, response) => {
+  const body = request.body
+  const updateBlog = {
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
   }
+  const result = await Blog.findByIdAndUpdate(request.params.id, updateBlog, {
+    new: true,
   })
+  response.json(result)
+})
 
 blogsRouter.put('/:id', (request, response, next) => {
   const body = request.body
